Avoid problem when logged in or not
Being logged off, and clicking in links to home.php?blabla... will bring
you to a page to log in.
Being logged in, and clicking on links to index.php?blabla... will bring
you to the good page, but will log you off in the process.

Added a (cheap) way to have both type of links available depending on
wether or not you are a logged in user, to be able to read the news
articles or change pages in News module with the appropriate destination
page (index.php for not logged in people, and home.php for logged in
users).