Couple of fixes
If &page is empty or has a period the script halts execution, and if a
value was given for a page that didn't exist we'd see no content.

In the circumstances above, it'll load the default index page now.
Additionally, the written by: section on add.php was changed so it'll
autofill the user's login name.